changing OneToOne relationship to OnteToMany
JPA Persistence spec 2.0 quote at 2.9:
"The inverse side of a bidirectional relationship must refer to its
owning side by use of the mappedBy element of the OneToOne, OneToMany,
or ManyToMany annotation."

3.2.4
Bidirectional relationships between managed entities will be persisted
based on references held by the owning side of the relationship.

3.2.4 states it later:
It is the developer’s responsibility to keep the in-memory references
held on the owning side and those held on the inverse side consistent
with each other when they change.